Refactor `initializers` logic
