+++ src/extension.ts
@@ -3,7 +3,7 @@
 import * as vscode from 'vscode';
 
 // Initializers
-import * as initialize from './initializers';
+import initialize from './initializers';
 
 // Commands
 import { commands } from './commands';
@@ -20,8 +20,8 @@ export async function activate(context: vscode.ExtensionContext) {
 	// Show Information Message when the extension is activated
 	vscode.window.showInformationMessage("BG3 Modding Extension Activated!");
 
-	// Initialize Localization FS Watcher
-	initialize.localizationFSWatcher();
+	// Initialize Localization Contributions
+	initialize.localization();
 
 	// Register all the commands and providers, and subscribe to their disposables
 	context.subscriptions.push(
