+++ src/initializers/index.ts
@@ -1,0 +1,10 @@
+// Library
+import { initializeLocalization } from './localization';
+
+// ============
+// INITIALIZERS
+// ============
+
+export default {
+    localization: initializeLocalization
+};
